pyproject.toml
[project]
name = "bdmc"
version = "0.1.0"
description = "An api wrapper lib designed for the uptech BDMC divers"
authors = [
    {name = "Whth", email = "88489697+Whth@users.noreply.github.com"},
]
dependencies = [
    "pyserial>=3.5",
    "coloredlogs>=15.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
